+++ android/src/main/java/ee/forgr/capacitor_updater/CapacitorUpdaterPlugin.java
@@ -226,6 +226,7 @@ public class CapacitorUpdaterPlugin extends Plugin {
     boolean resetWhenUpdate =
       this.getConfig().getBoolean("resetWhenUpdate", true);
 
+    this.implementation.autoReset();
     if (resetWhenUpdate) {
       this.cleanupObsoleteVersions();
     }
@@ -375,7 +376,10 @@ public class CapacitorUpdaterPlugin extends Plugin {
   @PluginMethod
   public void setUpdateUrl(final PluginCall call) {
     if (!this.getConfig().getBoolean("allowModifyUrl", false)) {
-      Log.e(CapacitorUpdater.TAG, "setUpdateUrl not allowed");
+      Log.e(
+        CapacitorUpdater.TAG,
+        "setUpdateUrl not allowed set allowModifyUrl in your config to true to allow it"
+      );
       call.reject("setUpdateUrl not allowed");
       return;
     }
@@ -392,7 +396,10 @@ public class CapacitorUpdaterPlugin extends Plugin {
   @PluginMethod
   public void setStatsUrl(final PluginCall call) {
     if (!this.getConfig().getBoolean("allowModifyUrl", false)) {
-      Log.e(CapacitorUpdater.TAG, "setStatsUrl not allowed");
+      Log.e(
+        CapacitorUpdater.TAG,
+        "setStatsUrl not allowed set allowModifyUrl in your config to true to allow it"
+      );
       call.reject("setStatsUrl not allowed");
       return;
     }
@@ -409,7 +416,10 @@ public class CapacitorUpdaterPlugin extends Plugin {
   @PluginMethod
   public void setChannelUrl(final PluginCall call) {
     if (!this.getConfig().getBoolean("allowModifyUrl", false)) {
-      Log.e(CapacitorUpdater.TAG, "setChannelUrl not allowed");
+      Log.e(
+        CapacitorUpdater.TAG,
+        "setChannelUrl not allowed set allowModifyUrl in your config to true to allow it"
+      );
       call.reject("setChannelUrl not allowed");
       return;
     }