+++ ios/Plugin/CapacitorUpdaterPlugin.swift
@@ -86,6 +86,7 @@ public class CapacitorUpdaterPlugin: CAPPlugin {
         implementation.statsUrl = getConfig().getString("statsUrl", CapacitorUpdaterPlugin.statsUrlDefault)!
         implementation.channelUrl = getConfig().getString("channelUrl", CapacitorUpdaterPlugin.channelUrlDefault)!
         implementation.defaultChannel = getConfig().getString("defaultChannel", "")!
+        self.implementation.autoReset()
         if resetWhenUpdate {
             self.cleanupObsoleteVersions()
         }
@@ -148,9 +149,9 @@ public class CapacitorUpdaterPlugin: CAPPlugin {
     }
 
     @objc func setUpdateUrl(_ call: CAPPluginCall) {
-        if !call.getBool("allowModifyUrl", false) {
+        if !getConfig().getBoolean("allowModifyUrl", false) {
             print("\(self.implementation.TAG) setUpdateUrl called without allowModifyUrl")
-            call.reject("setUpdateUrl called without allowModifyUrl")
+            call.reject("setUpdateUrl called without allowModifyUrl set allowModifyUrl in your config to true to allow it")
             return
         }
         guard let url = call.getString("url") else {
@@ -163,9 +164,9 @@ public class CapacitorUpdaterPlugin: CAPPlugin {
     }
 
     @objc func setStatsUrl(_ call: CAPPluginCall) {
-        if !call.getBool("allowModifyUrl", false) {
+        if !getConfig().getBoolean("allowModifyUrl", false) {
             print("\(self.implementation.TAG) setStatsUrl called without allowModifyUrl")
-            call.reject("setStatsUrl called without allowModifyUrl")
+            call.reject("setStatsUrl called without allowModifyUrl set allowModifyUrl in your config to true to allow it")
             return
         }
         guard let url = call.getString("url") else {
@@ -178,9 +179,9 @@ public class CapacitorUpdaterPlugin: CAPPlugin {
     }
 
     @objc func setChannelUrl(_ call: CAPPluginCall) {
-        if !call.getBool("allowModifyUrl", false) {
+        if !getConfig().getBoolean("allowModifyUrl", false) {
             print("\(self.implementation.TAG) setChannelUrl called without allowModifyUrl")
-            call.reject("setChannelUrl called without allowModifyUrl")
+            call.reject("setChannelUrl called without allowModifyUrl set allowModifyUrl in your config to true to allow it")
             return
         }
         guard let url = call.getString("url") else {